# Small speedup in AVX2/AVX512