--- a/avo/avx512.go
+++ b/avo/avx512.go
@@ -15,7 +15,6 @@ func AVX512() {
 	}
 
 	{
-
 		TEXT("accumAVX512", NOSPLIT, "func(acc *[8]uint64, data, key *byte, len uint64)")
 
 		acc := Mem{Base: Load(Param("acc"), GP64())}
@@ -23,11 +22,13 @@ func AVX512() {
 		key := Mem{Base: Load(Param("key"), GP64())}
 		plen := Load(Param("len"), GP64())
 		prime := ZMM()
-		a := ZMM()
+		// a[0] contains merged values
+		// a[1] is a temporary accumulator used for the accum_large loop.
+		a := [2]VecVirtual{ZMM(), ZMM()}
 
 		Label("load")
 		{
-			VMOVDQU64(acc.Offset(0x00), a)
+			VMOVDQU64(acc.Offset(0x00), a[0])
 			VMOVDQU64(primeData, prime)
 		}
 		// Load key at 8 byte offsets in the order we use it.
@@ -51,10 +52,11 @@ func AVX512() {
 			JLE(LabelRef("accum"))
 
 			for i := 0; i < 16; i++ {
-				avx512accum(data, a, 64*i, keyReg[i], true)
+				avx512accumSplit(data, a, 64*i, keyReg[i], true, i == 0)
 			}
+			VPADDQ(a[0], a[1], a[0])
 			advance(16)
-			avx512scramble(prime, keyReg[16], a)
+			avx512scramble(prime, keyReg[16], a[0])
 
 			JMP(LabelRef("accum_large"))
 		}
@@ -66,7 +68,7 @@ func AVX512() {
 				CMPQ(plen, Imm(64))
 				JLE(LabelRef("finalize"))
 
-				avx512accum(data, a, 0, keyReg[i], false)
+				avx512accum(data, a[0], 0, keyReg[i], false)
 				advance(1)
 			}
 		}
@@ -79,12 +81,12 @@ func AVX512() {
 			SUBQ(Imm(64), data.Base)
 			ADDQ(plen, data.Base)
 
-			avx512accum(data, a, 0, key121, false)
+			avx512accum(data, a[0], 0, key121, false)
 		}
 
 		Label("return")
 		{
-			VMOVDQU64(a, acc.Offset(0x00))
+			VMOVDQU64(a[0], acc.Offset(0x00))
 			VZEROUPPER()
 			RET()
 		}
@@ -118,6 +120,26 @@ func avx512accum(data Mem, a VecVirtual, doff int, key VecVirtual, prefetch bool
 	VPSHUFD(Imm(49), y1, y2)
 	VPMULUDQ(y1, y2, y1)
 	VPSHUFD(Imm(78), y0, y0)
-	VPADDQ(a, y0, a)
 	VPADDQ(a, y1, a)
+	VPADDQ(a, y0, a)
+}
+
+func avx512accumSplit(data Mem, a [2]VecVirtual, doff int, key VecVirtual, prefetch, initA1 bool) {
+	y0, y1, y2 := ZMM(), ZMM(), ZMM()
+	VMOVDQU64(data.Offset(doff+0x00), y0)
+	if initA1 {
+		y1 = a[1]
+	}
+	if prefetch {
+		// Prefetch once per cacheline (64 bytes), 8 iterations ahead.
+		PREFETCHT0(data.Offset(doff + 1024))
+	}
+	VPXORD(key, y0, y1)
+	VPSHUFD(Imm(49), y1, y2)
+	VPMULUDQ(y1, y2, y1)
+	VPSHUFD(Imm(78), y0, y0)
+	if !initA1 {
+		VPADDQ(a[1], y1, a[1])
+	}
+	VPADDQ(a[0], y0, a[0])
 }