… accum_vector_avx512_amd64.s
… accum_vector_avx_amd64.s