* Accumulate in separate registers in large loop to reduce latency.
* Load (reused) keys in unused AVX2 registers.
* Increase prefetch length in AVX2.

Benchmarks are a bit noisy from run to run, probably allocation alignment, but the overall trend appears positive.

```
name                                old time/op    new time/op    delta
Fixed128/1024-AVX2/default-32         28.9ns ± 1%    29.2ns ± 1%   +1.04%         (p=0.006 n=10+9)
Fixed128/1024-AVX2/seed-32            42.9ns ± 1%    42.8ns ± 1%     ~            (p=0.138 n=9+10)
Fixed128/8192-AVX2/default-32          137ns ± 1%     134ns ± 1%   -2.12%         (p=0.000 n=10+9)
Fixed128/8192-AVX2/seed-32             160ns ± 1%     160ns ± 1%     ~            (p=0.674 n=10+9)
Fixed128/102400-AVX2/default-32       1.56µs ± 1%    1.51µs ± 1%   -3.09%        (p=0.000 n=10+10)
Fixed128/102400-AVX2/seed-32          1.58µs ± 1%    1.53µs ± 0%   -3.12%        (p=0.000 n=10+10)
Fixed128/1024000-AVX2/default-32      15.4µs ± 1%    16.1µs ± 1%   +4.84%        (p=0.000 n=10+10)
Fixed128/1024000-AVX2/seed-32         15.6µs ± 1%    16.2µs ± 1%   +3.98%        (p=0.000 n=10+10)
Fixed128/10240000-AVX2/default-32      164µs ± 1%     157µs ± 0%   -4.23%        (p=0.000 n=10+10)
Fixed128/10240000-AVX2/seed-32         168µs ± 2%     158µs ± 1%   -6.16%        (p=0.000 n=10+10)
Fixed128/102400000-AVX2/default-32    3.91ms ± 1%    3.79ms ± 1%   -2.86%         (p=0.000 n=9+10)
Fixed128/102400000-AVX2/seed-32       3.92ms ± 1%    3.79ms ± 0%   -3.22%        (p=0.000 n=10+10)
Fixed64/1024-AVX2/default-32          26.0ns ± 1%    26.4ns ± 1%   +1.47%        (p=0.001 n=10+10)
Fixed64/1024-AVX2/seed-32             38.0ns ± 1%    38.2ns ± 1%   +0.48%        (p=0.037 n=10+10)
Fixed64/8192-AVX2/default-32           134ns ± 1%     133ns ± 0%   -1.11%         (p=0.000 n=10+8)
Fixed64/8192-AVX2/seed-32              146ns ± 1%     143ns ± 0%   -1.91%        (p=0.000 n=10+10)
Fixed64/102400-AVX2/default-32        1.56µs ± 1%    1.51µs ± 0%   -3.03%        (p=0.000 n=10+10)
Fixed64/102400-AVX2/seed-32           1.56µs ± 1%    1.52µs ± 1%   -2.85%        (p=0.000 n=10+10)
Fixed64/1024000-AVX2/default-32       16.0µs ± 1%    15.6µs ± 1%   -2.89%        (p=0.000 n=10+10)
Fixed64/1024000-AVX2/seed-32          16.1µs ± 1%    15.6µs ± 1%   -3.28%        (p=0.000 n=10+10)
Fixed64/10240000-AVX2/default-32       165µs ± 1%     157µs ± 0%   -4.60%        (p=0.000 n=10+10)
Fixed64/10240000-AVX2/seed-32          165µs ± 0%     157µs ± 0%   -4.71%         (p=0.000 n=9+10)
Fixed64/102400000-AVX2/default-32     3.91ms ± 1%    3.80ms ± 1%   -2.95%        (p=0.000 n=10+10)
Fixed64/102400000-AVX2/seed-32        3.90ms ± 1%    3.80ms ± 0%   -2.57%         (p=0.000 n=9+10)
Hasher64/16/avx2/plain-32             10.6ns ± 1%    10.6ns ± 1%   +0.58%        (p=0.012 n=10+10)
Hasher64/16/avx2/seed-32              10.6ns ± 1%    10.7ns ± 1%     ~           (p=0.069 n=10+10)
Hasher64/64/avx2/plain-32             13.3ns ± 1%    13.3ns ± 0%     ~           (p=0.867 n=10+10)
Hasher64/64/avx2/seed-32              13.4ns ± 0%    13.4ns ± 0%     ~             (p=0.093 n=8+9)
Hasher64/256/avx2/plain-32            25.2ns ± 1%    25.3ns ± 1%   +0.63%         (p=0.010 n=9+10)
Hasher64/256/avx2/seed-32             41.1ns ± 5%    37.8ns ± 1%   -8.08%        (p=0.000 n=10+10)
Hasher64/1024/avx2/plain-32           41.2ns ± 1%    41.4ns ± 2%     ~           (p=0.305 n=10+10)
Hasher64/1024/avx2/seed-32            53.8ns ± 1%    54.1ns ± 1%   +0.48%         (p=0.045 n=9+10)
Hasher64/4096/avx2/plain-32           92.7ns ± 1%    91.3ns ± 0%   -1.49%          (p=0.000 n=8+7)
Hasher64/4096/avx2/seed-32             106ns ± 2%      92ns ± 1%  -13.00%        (p=0.000 n=10+10)
Hasher64/16384/avx2/plain-32           300ns ± 1%     293ns ± 1%   -2.29%        (p=0.000 n=10+10)
Hasher64/16384/avx2/seed-32            316ns ± 1%     291ns ± 1%   -7.79%        (p=0.000 n=10+10)
Hasher64/65536/avx2/plain-32          1.16µs ± 0%    1.16µs ± 1%   -0.38%         (p=0.007 n=9+10)
Hasher64/65536/avx2/seed-32           1.17µs ± 1%    1.15µs ± 1%   -2.01%        (p=0.000 n=10+10)
Hasher64/262144/avx2/plain-32         4.58µs ± 1%    4.54µs ± 0%   -0.92%        (p=0.000 n=10+10)
Hasher64/262144/avx2/seed-32          4.52µs ± 1%    4.49µs ± 0%   -0.62%         (p=0.003 n=10+9)
Hasher64/1048576/avx2/plain-32        19.2µs ± 1%    19.1µs ± 1%   -0.57%        (p=0.030 n=10+10)
Hasher64/1048576/avx2/seed-32         19.2µs ± 1%    19.2µs ± 0%     ~           (p=0.481 n=10+10)
Hasher64/4194304/avx2/plain-32        76.8µs ± 1%    76.4µs ± 0%   -0.57%        (p=0.003 n=10+10)
Hasher64/4194304/avx2/seed-32         77.1µs ± 1%    76.7µs ± 0%     ~           (p=0.280 n=10+10)
Hasher64/16777216/avx2/plain-32        687µs ±19%     552µs ± 3%  -19.61%        (p=0.001 n=10+10)
Hasher64/16777216/avx2/seed-32         519µs ±11%     543µs ± 3%   +4.60%         (p=0.028 n=10+9)
Hasher64/67108864/avx2/plain-32       3.15ms ± 1%    3.11ms ± 0%   -1.40%         (p=0.000 n=10+8)
Hasher64/67108864/avx2/seed-32        3.12ms ± 1%    3.10ms ± 1%   -0.68%        (p=0.002 n=10+10)
Hasher64/268435456/avx2/plain-32      12.6ms ± 1%    12.5ms ± 1%   -0.71%         (p=0.004 n=9+10)
Hasher64/268435456/avx2/seed-32       12.5ms ± 0%    12.4ms ± 1%   -1.10%          (p=0.024 n=3+6)

name                                old speed      new speed      delta
Fixed128/1024-AVX2/default-32       35.5GB/s ± 1%  35.1GB/s ± 1%   -1.04%         (p=0.006 n=10+9)
Fixed128/1024-AVX2/seed-32          23.8GB/s ± 0%  23.9GB/s ± 1%   +0.48%         (p=0.043 n=8+10)
Fixed128/8192-AVX2/default-32       59.9GB/s ± 1%  61.2GB/s ± 1%   +2.17%         (p=0.000 n=10+9)
Fixed128/8192-AVX2/seed-32          51.1GB/s ± 1%  51.1GB/s ± 1%     ~            (p=0.661 n=10+9)
Fixed128/102400-AVX2/default-32     65.7GB/s ± 1%  67.8GB/s ± 1%   +3.20%        (p=0.000 n=10+10)
Fixed128/102400-AVX2/seed-32        64.8GB/s ± 1%  66.9GB/s ± 0%   +3.23%        (p=0.000 n=10+10)
Fixed128/1024000-AVX2/default-32    66.6GB/s ± 1%  63.6GB/s ± 1%   -4.61%        (p=0.000 n=10+10)
Fixed128/1024000-AVX2/seed-32       65.8GB/s ± 1%  63.3GB/s ± 1%   -3.82%        (p=0.000 n=10+10)
Fixed128/10240000-AVX2/default-32   62.3GB/s ± 1%  65.0GB/s ± 0%   +4.42%        (p=0.000 n=10+10)
Fixed128/10240000-AVX2/seed-32      60.9GB/s ± 2%  64.9GB/s ± 1%   +6.55%        (p=0.000 n=10+10)
Fixed128/102400000-AVX2/default-32  26.2GB/s ± 1%  27.0GB/s ± 1%   +3.11%        (p=0.000 n=10+10)
Fixed128/102400000-AVX2/seed-32     26.1GB/s ± 1%  27.0GB/s ± 0%   +3.33%        (p=0.000 n=10+10)
Fixed64/1024-AVX2/default-32        39.4GB/s ± 1%  38.8GB/s ± 1%   -1.46%        (p=0.002 n=10+10)
Fixed64/1024-AVX2/seed-32           26.9GB/s ± 1%  26.8GB/s ± 1%   -0.48%        (p=0.043 n=10+10)
Fixed64/8192-AVX2/default-32        61.1GB/s ± 1%  61.8GB/s ± 0%   +1.10%         (p=0.000 n=10+8)
Fixed64/8192-AVX2/seed-32           56.1GB/s ± 1%  57.2GB/s ± 0%   +1.95%        (p=0.000 n=10+10)
Fixed64/102400-AVX2/default-32      65.8GB/s ± 1%  67.9GB/s ± 0%   +3.12%        (p=0.000 n=10+10)
Fixed64/102400-AVX2/seed-32         65.5GB/s ± 1%  67.5GB/s ± 1%   +2.94%        (p=0.000 n=10+10)
Fixed64/1024000-AVX2/default-32     63.8GB/s ± 1%  65.7GB/s ± 1%   +2.97%        (p=0.000 n=10+10)
Fixed64/1024000-AVX2/seed-32        63.5GB/s ± 1%  65.7GB/s ± 1%   +3.38%        (p=0.000 n=10+10)
Fixed64/10240000-AVX2/default-32    62.1GB/s ± 1%  65.1GB/s ± 0%   +4.81%        (p=0.000 n=10+10)
Fixed64/10240000-AVX2/seed-32       62.1GB/s ± 0%  65.1GB/s ± 0%   +4.94%         (p=0.000 n=9+10)
Fixed64/102400000-AVX2/default-32   26.2GB/s ± 1%  27.0GB/s ± 1%   +3.04%        (p=0.000 n=10+10)
Fixed64/102400000-AVX2/seed-32      26.3GB/s ± 1%  26.9GB/s ± 0%   +2.64%         (p=0.000 n=9+10)
Hasher64/16/avx2/plain-32           1.52GB/s ± 1%  1.51GB/s ± 1%   -0.56%        (p=0.019 n=10+10)
Hasher64/16/avx2/seed-32            1.51GB/s ± 1%  1.50GB/s ± 1%     ~           (p=0.060 n=10+10)
Hasher64/64/avx2/plain-32           4.81GB/s ± 1%  4.81GB/s ± 0%     ~           (p=0.912 n=10+10)
Hasher64/64/avx2/seed-32            4.77GB/s ± 1%  4.78GB/s ± 0%     ~            (p=0.075 n=10+9)
Hasher64/256/avx2/plain-32          10.2GB/s ± 1%  10.1GB/s ± 1%   -0.62%         (p=0.013 n=9+10)
Hasher64/256/avx2/seed-32           6.24GB/s ± 5%  6.78GB/s ± 1%   +8.67%        (p=0.000 n=10+10)
Hasher64/1024/avx2/plain-32         24.8GB/s ± 1%  24.7GB/s ± 2%     ~           (p=0.315 n=10+10)
Hasher64/1024/avx2/seed-32          19.0GB/s ± 1%  18.9GB/s ± 1%     ~            (p=0.053 n=9+10)
Hasher64/4096/avx2/plain-32         44.2GB/s ± 1%  44.9GB/s ± 0%   +1.59%          (p=0.001 n=8+6)
Hasher64/4096/avx2/seed-32          38.6GB/s ± 3%  44.4GB/s ± 1%  +14.91%        (p=0.000 n=10+10)
Hasher64/16384/avx2/plain-32        54.7GB/s ± 1%  55.9GB/s ± 1%   +2.34%        (p=0.000 n=10+10)
Hasher64/16384/avx2/seed-32         51.8GB/s ± 1%  56.2GB/s ± 1%   +8.44%        (p=0.000 n=10+10)
Hasher64/65536/avx2/plain-32        56.3GB/s ± 0%  56.5GB/s ± 1%   +0.38%         (p=0.008 n=9+10)
Hasher64/65536/avx2/seed-32         56.0GB/s ± 1%  57.1GB/s ± 1%   +2.03%        (p=0.000 n=10+10)
Hasher64/262144/avx2/plain-32       57.3GB/s ± 1%  57.8GB/s ± 0%   +0.93%        (p=0.000 n=10+10)
Hasher64/262144/avx2/seed-32        58.0GB/s ± 1%  58.4GB/s ± 0%   +0.62%         (p=0.003 n=10+9)
Hasher64/1048576/avx2/plain-32      54.5GB/s ± 1%  54.8GB/s ± 1%   +0.57%        (p=0.029 n=10+10)
Hasher64/1048576/avx2/seed-32       54.6GB/s ± 1%  54.6GB/s ± 0%     ~           (p=0.481 n=10+10)
Hasher64/4194304/avx2/plain-32      54.6GB/s ± 1%  54.9GB/s ± 0%   +0.57%        (p=0.003 n=10+10)
Hasher64/4194304/avx2/seed-32       54.4GB/s ± 1%  54.7GB/s ± 0%     ~           (p=0.280 n=10+10)
Hasher64/16777216/avx2/plain-32     24.9GB/s ±21%  30.4GB/s ± 3%  +22.34%        (p=0.001 n=10+10)
Hasher64/16777216/avx2/seed-32      32.4GB/s ±10%  30.9GB/s ± 3%   -4.65%         (p=0.028 n=10+9)
Hasher64/67108864/avx2/plain-32     21.3GB/s ± 1%  21.6GB/s ± 0%   +1.42%         (p=0.000 n=10+8)
Hasher64/67108864/avx2/seed-32      21.5GB/s ± 1%  21.6GB/s ± 1%   +0.69%        (p=0.002 n=10+10)
Hasher64/268435456/avx2/plain-32    21.4GB/s ± 1%  21.5GB/s ± 1%   +0.71%         (p=0.004 n=9+10)
Hasher64/268435456/avx2/seed-32     21.4GB/s ± 0%  21.7GB/s ± 1%   +1.11%          (p=0.024 n=3+6)
```

My Zen 2 has memory -> register aliasing - this is not present on Intel, so they preloaded keys should help other platforms more.

I will have to run tests on AVX512. I will get back with whether they are good if CI doesn't pick it up.

Edit: It was picked up by CI:

=== RUN   TestVectorCompat
    compat_vector_test.go:[30](https://github.com/zeebo/xxh3/actions/runs/4850815085/jobs/8644067832?pr=22#step:4:31): avx512: true
    compat_vector_test.go:[31](https://github.com/zeebo/xxh3/actions/runs/4850815085/jobs/8644067832?pr=22#step:4:32): avx2: true
    compat_vector_test.go:32: sse2: true